final final final version

+++ main.c
@@ -120,6 +120,7 @@ int CopyFileNormal(char *sourcePath, char *destinationPath)
     close(copyFromFile);
     close(copyToFile);
     free(buffer);
+    return 0;
 
 }
 // Kopiowanie pliku z katalogu 1 do katalogu 2 za pomoca mmap/write
@@ -143,6 +144,7 @@ int CopyFileMmap(char *sourcePath, char *destinationPath)
     syslog(LOG_NOTICE, "COPYING [mmap] %s TO %s",sourcePath,destinationPath);
     close(copyFromFile);
     close(copyToFile);
+    return 0;
 
 }
 // Usuwanie pliku / katalogu
@@ -165,12 +167,20 @@ int DeleteEntry(char *relativePath)
                     return 1; //exit(EXIT_FAILURE);
                 }
 
+                int dirWithLinks;
+
                 while((entry = readdir(dir)) != NULL)
                 {
                     if(strcmp(entry->d_name, ".") == 0 || strcmp(entry->d_name, "..") == 0)
                         continue;
 
-                    DeleteEntry(AddFileNameToDirPath(relativePath, entry->d_name));
+                    dirWithLinks = DeleteEntry(AddFileNameToDirPath(relativePath, entry->d_name));
+                }
+
+                if (dirWithLinks)
+                {
+                    syslog(LOG_ERR, "CAN'T DELETE DIR-CONTAINS LINKS: %s", relativePath);
+                    return 1; // exit(EXIT_FAILURE);
                 }
 
                 if (unlinkat(AT_FDCWD, relativePath, AT_REMOVEDIR))
@@ -188,7 +198,7 @@ int DeleteEntry(char *relativePath)
 
         syslog(LOG_NOTICE, "DELETED DIRECTORY: %s\n", relativePath);
     }
-    else
+    else if(isRegularFile(relativePath))
     {
         if (unlink(relativePath))
         {
@@ -197,6 +207,10 @@ int DeleteEntry(char *relativePath)
         }
 
         syslog(LOG_NOTICE, "DELETED FILE: %s\n", relativePath);
+    }
+    else
+    {
+        return 1;
     }
 
     return 0; //poprawnie usunięto wpis
@@ -486,8 +500,7 @@ void SyncDelete(char *sourceDirPath, char *destDirPath)
 
         //pomijanie twardych dowiązań do obecnego i nadrzędnego katalogu
         if ((strcmp(entryDest->d_name, ".") == 0 || strcmp(entryDest->d_name, "..") == 0)
-            || (allowRecursion == false && isDirectory(destEntryPath)) //jeżeli jest katalogiem
-            || (isRegularFile(destEntryPath) == 0 && isDirectory(destEntryPath) == 0)) //nie jest ani katalogiem ani plikiem
+            || (allowRecursion == false && isRegularFile(destEntryPath) == 0)) //jeżeli nie jest plikiem
         {
             free(destEntryPath);
             continue;
@@ -522,8 +535,6 @@ void SyncDelete(char *sourceDirPath, char *destDirPath)
 
                 break;
             }
-
-            free(sourceEntryPath);
         }
 
         if(allowDelete)
@@ -531,6 +542,7 @@ void SyncDelete(char *sourceDirPath, char *destDirPath)
             DeleteEntry(destEntryPath);
         }
 
+        free(sourceEntryPath);
         free(destEntryPath);
 
         closedir(dir_source);
